refactor editor mapitems
Implement `CEditorComponent` for `CEditorMap` and all
`CLayer` types. Also register Zoom as subcomponent and
remove `OnMapLoad` for `CProofMode`.
