+++ src/game/editor/map_grid.cpp
@@ -1,6 +1,6 @@
 #include "map_grid.h"
 
-#include <engine/keys.h>
+#include <game/editor/mapitems/layer_group.h>
 
 #include "editor.h"
