+++ src/game/editor/mapitems/layer_game.cpp
@@ -1,6 +1,11 @@
 /* (c) Magnus Auvinen. See licence.txt in the root of the distribution for more information. */
 /* If you are missing that file, acquire a complete release at teeworlds.com.                */
-#include "editor.h"
+#include "layer_game.h"
+
+#include <game/editor/editor.h>
+
+#include "layer_front.h"
+#include "layer_group.h"
 
 CLayerGame::CLayerGame(int w, int h) :
 	CLayerTiles(w, h)
@@ -13,7 +18,7 @@ CLayerGame::~CLayerGame() = default;
 
 CTile CLayerGame::GetTile(int x, int y)
 {
-	if(m_pEditor->m_Map.m_pFrontLayer && m_pEditor->m_Map.m_pFrontLayer->GetTile(x, y).m_Index == TILE_THROUGH_CUT)
+	if(Editor()->m_Map.m_pFrontLayer && Editor()->m_Map.m_pFrontLayer->GetTile(x, y).m_Index == TILE_THROUGH_CUT)
 	{
 		CTile through_cut = {TILE_THROUGH_CUT};
 		return through_cut;
@@ -26,27 +31,27 @@ CTile CLayerGame::GetTile(int x, int y)
 
 void CLayerGame::SetTile(int x, int y, CTile Tile)
 {
-	if(Tile.m_Index == TILE_THROUGH_CUT && m_pEditor->m_SelectEntitiesImage == "DDNet")
+	if(Tile.m_Index == TILE_THROUGH_CUT && Editor()->m_SelectEntitiesImage == "DDNet")
 	{
-		if(!m_pEditor->m_Map.m_pFrontLayer)
+		if(!Editor()->m_Map.m_pFrontLayer)
 		{
 			std::shared_ptr<CLayer> pLayerFront = std::make_shared<CLayerFront>(m_Width, m_Height);
-			m_pEditor->m_Map.MakeFrontLayer(pLayerFront);
-			m_pEditor->m_Map.m_pGameGroup->AddLayer(pLayerFront);
+			Editor()->m_Map.MakeFrontLayer(pLayerFront);
+			Editor()->m_Map.m_pGameGroup->AddLayer(pLayerFront);
 		}
 		CTile nohook = {TILE_NOHOOK};
 		CLayerTiles::SetTile(x, y, nohook);
 		CTile through_cut = {TILE_THROUGH_CUT};
-		m_pEditor->m_Map.m_pFrontLayer->CLayerTiles::SetTile(x, y, through_cut); // NOLINT(bugprone-parent-virtual-call)
+		Editor()->m_Map.m_pFrontLayer->CLayerTiles::SetTile(x, y, through_cut); // NOLINT(bugprone-parent-virtual-call)
 	}
 	else
 	{
-		if(m_pEditor->m_SelectEntitiesImage == "DDNet" && m_pEditor->m_Map.m_pFrontLayer && m_pEditor->m_Map.m_pFrontLayer->GetTile(x, y).m_Index == TILE_THROUGH_CUT)
+		if(Editor()->m_SelectEntitiesImage == "DDNet" && Editor()->m_Map.m_pFrontLayer && Editor()->m_Map.m_pFrontLayer->GetTile(x, y).m_Index == TILE_THROUGH_CUT)
 		{
 			CTile air = {TILE_AIR};
-			m_pEditor->m_Map.m_pFrontLayer->CLayerTiles::SetTile(x, y, air); // NOLINT(bugprone-parent-virtual-call)
+			Editor()->m_Map.m_pFrontLayer->CLayerTiles::SetTile(x, y, air); // NOLINT(bugprone-parent-virtual-call)
 		}
-		if(m_pEditor->m_AllowPlaceUnusedTiles || IsValidGameTile(Tile.m_Index))
+		if(Editor()->m_AllowPlaceUnusedTiles || IsValidGameTile(Tile.m_Index))
 		{
 			CLayerTiles::SetTile(x, y, Tile);
 		}
@@ -54,11 +59,11 @@ void CLayerGame::SetTile(int x, int y, CTile Tile)
 		{
 			CTile air = {TILE_AIR};
 			CLayerTiles::SetTile(x, y, air);
-			if(!m_pEditor->m_PreventUnusedTilesWasWarned)
+			if(!Editor()->m_PreventUnusedTilesWasWarned)
 			{
-				m_pEditor->m_PopupEventType = CEditor::POPEVENT_PREVENTUNUSEDTILES;
-				m_pEditor->m_PopupEventActivated = true;
-				m_pEditor->m_PreventUnusedTilesWasWarned = true;
+				Editor()->m_PopupEventType = CEditor::POPEVENT_PREVENTUNUSEDTILES;
+				Editor()->m_PopupEventActivated = true;
+				Editor()->m_PreventUnusedTilesWasWarned = true;
 			}
 		}
 	}
