+++ src/game/editor/mapitems/layer_sounds.h
@@ -1,0 +1,33 @@
+#ifndef GAME_EDITOR_MAPITEMS_LAYER_SOUNDS_H
+#define GAME_EDITOR_MAPITEMS_LAYER_SOUNDS_H
+
+#include "layer.h"
+
+struct CSoundSource;
+
+class CLayerSounds : public CLayer
+{
+public:
+	CLayerSounds();
+	CLayerSounds(const CLayerSounds &Other);
+	~CLayerSounds();
+
+	void Render(bool Tileset = false) override;
+	CSoundSource *NewSource(int x, int y);
+
+	void BrushSelecting(CUIRect Rect) override;
+	int BrushGrab(const std::shared_ptr<CLayerGroup> &pBrush, CUIRect Rect) override;
+	void BrushPlace(const std::shared_ptr<CLayer> &pBrush, float wx, float wy) override;
+
+	CUI::EPopupMenuFunctionResult RenderProperties(CUIRect *pToolbox) override;
+
+	void ModifyEnvelopeIndex(const FIndexModifyFunction &pfnFunc) override;
+	void ModifySoundIndex(const FIndexModifyFunction &pfnFunc) override;
+
+	std::shared_ptr<CLayer> Duplicate() const override;
+
+	int m_Sound;
+	std::vector<CSoundSource> m_vSources;
+};
+
+#endif
